backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary
Pillow==10.4.0
python-decouple==3.8
python-magic==0.4.27
hypothesis==6.88.1
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
requests==2.31.0
PyPDF2==3.0.1